requirements.txt
click
langchain
langchain-community
openai
python-dotenv
setuptools
upstash_redis